Set up test framework for index.js
Ember's testing system is all focused on testing the browser app.
I pulled in `qunit` from NPM, so as to use the same test framework
as is used in the Ember side, but it looks like it pulls in
an ancient version of QUnit. It's passable, though, and enough
yak fur has been shed that I'm continuing with this.